+++ MachineLearning.Samples/GlobalUsings.cs
@@ -13,4 +13,4 @@ global using MachineLearning.Training.Optimization.Adam;
 global using MachineLearning.Training.Optimization.AdamW;
 global using MachineLearning.Training.Optimization.Nadam;
 global using MachineLearning.Training.Optimization.SGDMomentum;
-global using Weight = double;
+global using Weight = float;
